# api/openapi.yaml
openapi: 3.0.3
info:
  title: Finenumbers SMS Service
  version: 0.1.0
  description: |
    Three surfaces, one process:
    - `/admin/v1` cookie admin (host admin.{domain})
    - `/client/v1` cookie client (host client.{domain})
    - `/v1` API keys (host api.{domain})
    - `/internal/runexis` provider callbacks
    - `/internal/smsc/callback` SMSC HLR/Ping callbacks
servers:
  - url: /
paths:
...
  /admin/v1/auth/login:
    post:
      tags: [admin]
      operationId: adminLogin
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session cookie set
        "401":
          description: Invalid credentials
        "429":
          description: Rate limited
  /admin/v1/auth/logout:
    post:
      tags: [admin]
      operationId: adminLogout
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "204":
          description: Cookie cleared
  /admin/v1/auth/me:
    get:
      tags: [admin]
      operationId: adminMe
      security: [{adminCookie: []}]
      responses:
        "200":
          description: Current admin
  /admin/v1/clients:
    get:
      tags: [admin]
      operationId: adminListClients
      security: [{adminCookie: []}]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Client list
    post:
      tags: [admin]
      operationId: adminCreateClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client and owner created
  /admin/v1/clients/{clientID}:
    get:
      tags: [admin]
      operationId: adminGetClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          description: Client with users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      tags: [admin]
      operationId: adminPatchClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchClientRequest"
      responses:
        "200":
          description: Updated client
    delete:
      tags: [admin]
      operationId: adminDeleteClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      responses:
        "204":
          description: Access revoked and tenant data wiped; numbers returned to inventory. Large histories finish on the worker.
  /admin/v1/clients/{clientID}/suspend:
    post:
      tags: [admin]
      operationId: adminSuspendClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          description: Suspended
        "409":
          description: Wrong status
  /admin/v1/clients/{clientID}/activate:
    post:
      tags: [admin]
      operationId: adminActivateClient
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          description: Activated
  /admin/v1/clients/{clientID}/billing:
    get:
      tags: [admin]
      operationId: adminGetClientBilling
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          description: Wallet, ledger, assigned tariffs
  /admin/v1/clients/{clientID}/tariff:
    post:
      tags: [admin]
      operationId: adminAssignClientTariff
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product, tariff_plan_id]
              properties:
                product:
                  type: string
                  enum: [sms_domestic, sms_international, hlr, silent_sms]
                tariff_plan_id:
                  type: string
                  format: uuid
                price_override:
                  type: string
      responses:
        "200":
          description: Assigned
        "400":
          description: Invalid product or plan
        "404":
          description: Client or plan not found
  /admin/v1/clients/{clientID}/tariff/{product}:
    delete:
      tags: [admin]
      operationId: adminUnassignClientTariff
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - name: product
          in: path
          required: true
          schema:
            type: string
            enum: [sms_domestic, sms_international, hlr, silent_sms]
      responses:
        "200":
          description: Assignment removed; client falls back to no sell price for the product
        "404":
          description: No assignment for this product
        "400":
          description: Invalid product or plan
  /admin/v1/clients/{clientID}/users:
    post:
      tags: [admin]
      operationId: adminCreateClientUser
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
                  minLength: 10
      responses:
        "201":
          description: Owner created
        "409":
          description: Email taken or user limit
  /admin/v1/clients/{clientID}/users/{userID}/name:
    post:
      tags: [admin]
      operationId: adminUpdateClientUserName
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Name updated
        "404":
          description: User not on this client
  /admin/v1/clients/{clientID}/users/{userID}/password:
    post:
      tags: [admin]
      operationId: adminResetClientUserPassword
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password updated; that user's sessions revoked
        "404":
          description: User not on this client
  /admin/v1/clients/{clientID}/users/{userID}/disable:
    post:
      tags: [admin]
      operationId: adminDisableClientUser
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Disabled
        "409":
          description: Last active owner
  /admin/v1/clients/{clientID}/users/{userID}/enable:
    post:
      tags: [admin]
      operationId: adminEnableClientUser
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Enabled
  /admin/v1/clients/{clientID}/owner/password:
    post:
      tags: [admin]
      operationId: adminResetOwnerPassword
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password updated
  /admin/v1/clients/{clientID}/api-keys:
    get:
      tags: [admin]
      operationId: adminListAPIKeys
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/ClientID"
      responses:
        "200":
          description: API keys (no secrets)
        "404":
          description: Client not found
    post:
      tags: [admin]
      operationId: adminCreateAPIKey
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: Key created; token shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
        "400":
          description: Validation
        "404":
          description: Client not found
  /admin/v1/clients/{clientID}/api-keys/{keyID}/revoke:
    post:
      tags: [admin]
      operationId: adminRevokeAPIKey
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/ClientID"
        - $ref: "#/components/parameters/KeyID"
      responses:
        "200":
          description: Revoked
        "404":
          description: Not found
        "409":
          description: Already revoked
  /admin/v1/numbers:
    get:
      tags: [admin]
      operationId: adminListNumbers
      security: [{adminCookie: []}]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [inventory, assigned, disabled]
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: DEF inventory
  /admin/v1/numbers/upload:
    post:
      tags: [admin]
      operationId: adminUploadNumbers
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import report (UTF-8 or Windows-1251)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberUploadReport"
  /admin/v1/numbers/sync:
    post:
      tags: [admin]
      operationId: adminSyncNumbers
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "200":
          description: DIDAPI inventory sync report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberSyncReport"
        "409":
          description: Runexis not configured, or a sync is already in progress
        "502":
          description: Runexis request failed
  /admin/v1/numbers/{numberID}:
    get:
      tags: [admin]
      operationId: adminGetNumber
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/NumberID"
      responses:
        "200":
          description: DEF number
        "404":
          description: Not found
    patch:
      tags: [admin]
      operationId: adminPatchNumber
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/NumberID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchNumberRequest"
      responses:
        "200":
          description: Updated number
        "409":
          description: Assigned numbers must be unassigned before disable
  /admin/v1/numbers/{numberID}/assign:
    post:
      tags: [admin]
      operationId: adminAssignNumber
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/NumberID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignNumberRequest"
      responses:
        "200":
          description: Assigned; directions job enqueued
        "409":
          description: Not assignable
  /admin/v1/numbers/{numberID}/unassign:
    post:
      tags: [admin]
      operationId: adminUnassignNumber
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/NumberID"
      responses:
        "200":
          description: Returned to inventory
        "409":
          description: Not assigned
  /admin/v1/billing/overview:
    get:
      tags: [admin]
      operationId: adminBillingOverview
      security: [{adminCookie: []}]
      responses:
        "200":
          description: Platform SMS and wallet totals
  /admin/v1/tariffs:
    get:
      tags: [admin]
      operationId: adminListTariffs
      security: [{adminCookie: []}]
      responses:
        "200":
          description: Tariff catalog
    post:
      tags: [admin]
      operationId: adminCreateTariff
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "201":
          description: Created plan
  /admin/v1/settings:
    get:
      tags: [admin]
      operationId: adminGetSettings
      security: [{adminCookie: []}]
      responses:
        "200":
          description: System settings (no secrets)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    patch:
      tags: [admin]
      operationId: adminPatchSettings
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
  /admin/v1/settings/runexis/test:
    post:
      tags: [admin]
      operationId: adminTestRunexis
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "200":
          description: Login + /me succeeded. statistic_ok / settings_ok are extra probes and do not change this status.
          content:
            application/json:
              schema:
                type: object
                required: [ok, email, name, statistic_ok]
                properties:
                  ok: {type: boolean}
                  email: {type: string}
                  name: {type: string}
                  statistic_ok: {type: boolean}
                  statistic_error: {type: string}
                  settings_ok: {type: boolean}
                  settings_error: {type: string}
                  hook_url: {type: string, description: Global MO hook from GET /sms/settings (no dlr_url in DIDAPI)}
                  in: {type: boolean}
                  dom_out: {type: boolean}
                  int_out: {type: boolean}
        "409":
          description: Credentials missing or cannot be decrypted
        "502":
          description: DIDAPI rejected the request
  /admin/v1/settings/runexis/callbacks:
    post:
      tags: [admin]
      operationId: adminRegisterCallbacks
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ingress_token]
              properties:
                ingress_token:
                  type: string
      responses:
        "200":
          description: Global dlr-url and hook-url registered
        "400":
          description: Token does not match stored hash
        "409":
          description: callback_base_url is not set
  /admin/v1/callbacks:
    get:
      tags: [admin]
      operationId: adminListCallbacks
      security: [{adminCookie: []}]
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [dlr, mo]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Captured DLR/MO events (no token in path)
  /admin/v1/callbacks/{callbackID}:
    get:
      tags: [admin]
      operationId: adminGetCallback
      security: [{adminCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CallbackID"
      responses:
        "200":
          description: Raw captured callback for fixture work
        "404":
          description: Not found
  /admin/v1/logs:
    get:
      tags: [admin]
      operationId: adminListLogs
      security: [{adminCookie: []}]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: RFC3339 UTC. Default now-1h. Max span with to is 24h.
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: category
          in: query
          schema:
            type: string
            enum: [http, didapi, queue, ingress, audit]
        - name: level
          in: query
          schema:
            type: string
            enum: [info, warn, error]
        - name: q
          in: query
          schema:
            type: string
          description: ILIKE on summary, action, error (not detail)
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
        - name: request_id
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Operational events without detail JSON
  /admin/v1/logs/{logID}:
    get:
      tags: [admin]
      operationId: adminGetLog
      security: [{adminCookie: []}]
      parameters:
        - name: logID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event including redacted detail
        "404":
          description: Not found
...
components:
  parameters:
    CSRF:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        example: XMLHttpRequest
    ClientID:
      name: clientID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
  securitySchemes:
    adminCookie:
      type: apiKey
      in: cookie
      name: __Host-fn_admin_sid
    clientCookie:
      type: apiKey
      in: cookie
      name: __Host-fn_client_sid
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: fnk_live_{prefix}_{secret}
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CreateClientRequest:
      type: object
      required: [name, owner_name, owner_email, owner_password]
      properties:
        name:
          type: string
        owner_name:
          type: string
        owner_email:
          type: string
          format: email
        owner_password:
          type: string
          minLength: 10
...